data/requirements.txt
psycopg2-binary==2.9.5
black==22.12.0
python-dotenv==0.21.0
requests==2.28.2
